Be literal about the types used in TokenInteger and TokenFloat

+++ lib/flux-internal.h
@@ -48,12 +48,12 @@ typedef TokenString TokenKeyword;
 
 typedef struct {
   TokenHeader header;
-  long number;
+  int number;
 } TokenInteger;
 
 typedef struct {
   TokenHeader header;
-  double number;
+  float number;
 } TokenFloat;
 
 typedef struct {
